Modified Screen + added first concrete Viewable

+++ src/main/java/it/polimi/ingsw/view/cli/Viewable.java
@@ -2,7 +2,7 @@ package it.polimi.ingsw.view.cli;
 
 public interface Viewable {
 
-    public void update(byte[] update);
+    public void update(int[] update);
 
     public String toString();
 }
